[CLI] Move main to octobot
